Account and Navigation Restructure
