+++ src/components/App/index.js
@@ -21,22 +21,6 @@ import AccountPage from '../Account';
 import AdminPage from '../Admin';
 import OnboardingPage from '../Onboarding';
 
-// const App = () => (
-//   <Router>
-//     <div>
-//       <Navigation />
-//       <hr />
-//       <Route exact path={ROUTES.LANDING} component={LandingPage} />
-//       <Route path={ROUTES.SIGN_UP} component={SignUpPage}/>
-//       <Route path={ROUTES.SIGN_IN} component={SignInPage} />
-//       <Route path={ROUTES.PASSWORD_FORGET} component={PasswordForgetPage} />
-//       <Route path={ROUTES.HOME} component={HomePage} />
-//       <Route path={ROUTES.ACCOUNT} component={AccountPage} />
-//       <Route path={ROUTES.ADMIN} component={AdminPage} />
-//     </div>
-//   </Router>
-// );
-
 class App extends Component {
   constructor(props) {
     super(props);
@@ -85,7 +69,6 @@ class App extends Component {
       </Router>
     );
   }
-
 }
  
 //export default withAuthentication(App);
