+++ src/components/Navigation/NavigationBarIcon.js
@@ -1,0 +1,40 @@
+import React from 'react';
+import { Col, Row } from 'react-bootstrap';
+
+import NavigationIconAccount from './NavigationIconAccount';
+import NavigationIconChat from './NavigationIconChat';
+import NavigationIconFriends from './NavigationIconFriends';
+import NavigationIconHome from './NavigationIconHome';
+import NavigationIconLibrary from './NavigationIconLibrary';
+
+import '../Styles/styles.css';
+
+const NavigationBarIcon = () => (
+    <div>
+      <Col className="naviconcol">
+        <Row>
+          <Col></Col>
+          <Col></Col>
+          <Col>
+            <NavigationIconHome />
+          </ Col>
+          <Col>
+            <NavigationIconChat />
+          </Col>
+          <Col>
+            <NavigationIconLibrary />
+          </Col>
+          <Col>
+            <NavigationIconFriends />
+          </Col>
+          <Col>
+            <NavigationIconAccount />
+          </Col>
+          <Col></Col>
+          <Col></Col>
+        </Row>
+      </Col>
+    </div>
+);
+
+export default NavigationBarIcon;
